Reorder the files not to build test code when not needed
