+++ helpers_test.go
@@ -3,10 +3,12 @@ package manners
 import (
 	"bufio"
 	"crypto/tls"
+	"errors"
 	"io/ioutil"
 	"net"
 	"net/http"
 	"os"
+	"sync"
 	"testing"
 )
 
@@ -139,3 +141,94 @@ func (tf *tempFile) Unlink() {
 		tf.File = nil
 	}
 }
+
+type testWg struct {
+	sync.Mutex
+	count      int
+	waitCalled chan int
+}
+
+func newTestWg() *testWg {
+	return &testWg{
+		waitCalled: make(chan int, 1),
+	}
+}
+
+func (wg *testWg) Add(delta int) {
+	wg.Lock()
+	wg.count++
+	wg.Unlock()
+}
+
+func (wg *testWg) Done() {
+	wg.Lock()
+	wg.count--
+	wg.Unlock()
+}
+
+func (wg *testWg) Wait() {
+	wg.Lock()
+	wg.waitCalled <- wg.count
+	wg.Unlock()
+}
+
+type fakeConn struct {
+	net.Conn
+	closeCalled bool
+}
+
+func (c *fakeConn) Close() error {
+	c.closeCalled = true
+	return nil
+}
+
+type fakeListener struct {
+	acceptRelease chan bool
+	closeCalled   chan bool
+}
+
+func newFakeListener() *fakeListener { return &fakeListener{make(chan bool, 1), make(chan bool, 1)} }
+
+func (l *fakeListener) Addr() net.Addr {
+	addr, _ := net.ResolveTCPAddr("tcp", "127.0.0.1:8080")
+	return addr
+}
+
+func (l *fakeListener) Close() error {
+	l.closeCalled <- true
+	l.acceptRelease <- true
+	return nil
+}
+
+func (l *fakeListener) Accept() (net.Conn, error) {
+	<-l.acceptRelease
+	return nil, errors.New("connection closed")
+}
+
+// localhostCert is a PEM-encoded TLS cert with SAN IPs
+// "127.0.0.1" and "[::1]", expiring at the last second of 2049 (the end
+// of ASN.1 time).
+// generated from src/pkg/crypto/tls:
+// go run generate_cert.go  --rsa-bits 512 --host 127.0.0.1,::1,example.com --ca --start-date "Jan 1 00:00:00 1970" --duration=1000000h
+var (
+	localhostCert = []byte(`-----BEGIN CERTIFICATE-----
+MIIBdzCCASOgAwIBAgIBADALBgkqhkiG9w0BAQUwEjEQMA4GA1UEChMHQWNtZSBD
+bzAeFw03MDAxMDEwMDAwMDBaFw00OTEyMzEyMzU5NTlaMBIxEDAOBgNVBAoTB0Fj
+bWUgQ28wWjALBgkqhkiG9w0BAQEDSwAwSAJBAN55NcYKZeInyTuhcCwFMhDHCmwa
+IUSdtXdcbItRB/yfXGBhiex00IaLXQnSU+QZPRZWYqeTEbFSgihqi1PUDy8CAwEA
+AaNoMGYwDgYDVR0PAQH/BAQDAgCkMBMGA1UdJQQMMAoGCCsGAQUFBwMBMA8GA1Ud
+EwEB/wQFMAMBAf8wLgYDVR0RBCcwJYILZXhhbXBsZS5jb22HBH8AAAGHEAAAAAAA
+AAAAAAAAAAAAAAEwCwYJKoZIhvcNAQEFA0EAAoQn/ytgqpiLcZu9XKbCJsJcvkgk
+Se6AbGXgSlq+ZCEVo0qIwSgeBqmsJxUu7NCSOwVJLYNEBO2DtIxoYVk+MA==
+-----END CERTIFICATE-----`)
+
+	localhostKey = []byte(`-----BEGIN RSA PRIVATE KEY-----
+MIIBPAIBAAJBAN55NcYKZeInyTuhcCwFMhDHCmwaIUSdtXdcbItRB/yfXGBhiex0
+0IaLXQnSU+QZPRZWYqeTEbFSgihqi1PUDy8CAwEAAQJBAQdUx66rfh8sYsgfdcvV
+NoafYpnEcB5s4m/vSVe6SU7dCK6eYec9f9wpT353ljhDUHq3EbmE4foNzJngh35d
+AekCIQDhRQG5Li0Wj8TM4obOnnXUXf1jRv0UkzE9AHWLG5q3AwIhAPzSjpYUDjVW
+MCUXgckTpKCuGwbJk7424Nb8bLzf3kllAiA5mUBgjfr/WtFSJdWcPQ4Zt9KTMNKD
+EUO0ukpTwEIl6wIhAMbGqZK3zAAFdq8DD2jPx+UJXnh0rnOkZBzDtJ6/iN69AiEA
+1Aq8MJgTaYsDQWyU/hDq5YkDJc9e9DSCvUIzqxQWMQE=
+-----END RSA PRIVATE KEY-----`)
+)
